added check for temporary  log file before blindly trying to removing it to keep mon log clean

+++ supermicro/check_sm_soft.sh
@@ -66,5 +66,8 @@ ${DETAILS}
 
 $LOCATIONLINE
 "
-rm $LOGFILE
+if [ -e $logfile ]; then
+	rm $LOGFILE
+fi
+
 exit 0
